requirements.txt
Flask_SQLAlchemy==2.5.1
numpy==1.19.3
opencv_python==4.5.2.52
pandas==1.0.1
Flask==1.1.1
mysql_connector_repackaged==0.3.1
